+++ maplibre_gl/android/src/main/java/org/maplibre/maplibregl/MapLibreMapOptionsSink.kt
@@ -54,6 +54,8 @@ internal interface MapLibreMapOptionsSink {
 
     fun setLocationEngineProperties(locationEngineRequest: LocationEngineRequest)
 
+    fun setLocationSource(token: String)
+
     fun setForegroundLoadColor(loadColor: Int)
 
     fun setTranslucentTextureSurface(translucentTextureSurface: Boolean)