+++ maplibre_gl/ios/maplibre_gl/Sources/maplibre_gl/MapLibreMapOptionsSink.swift
@@ -23,5 +23,11 @@ protocol MapLibreMapOptionsSink {
     func setAttributionButtonPosition(position: MLNOrnamentPosition)
     func setAttributionButtonColor(color: Int)
     func setFeatureTapsTriggersMapClick(triggers: Bool)
-    func setLocationEngineProperties(enableHighAccuracy: Bool, distanceFilter: Double, intervalMs: Int, pulseWindowMs: Int)
+    func setLocationEngineProperties(
+        enableHighAccuracy: Bool,
+        distanceFilter: Double,
+        intervalMs: Int,
+        pulseWindowMs: Int
+    )
+    func setLocationSource(token: String)
 }